fix: don't display "undefined" when no profiles exists
If you've not created a profile yet and attempt to run a command that requires auth, you would see that the profile "undefined" could not be found. This change just updates the wording to indicate that a profile could not be found to use.

Relates to https://github.com/twilio/twilio-cli/issues/148